cheri: usb: Fix compile in drivers/usb/serial/
Mostly ioctl issues.

+++ drivers/usb/serial/ftdi_sio.c
@@ -2854,7 +2854,7 @@ static int ftdi_tiocmset(struct tty_struct *tty,
 }
 
 static int ftdi_ioctl(struct tty_struct *tty,
-					unsigned int cmd, unsigned long arg)
+					unsigned int cmd, uintptr_t arg)
 {
 	struct usb_serial_port *port = tty->driver_data;
 	void __user *argp = (void __user *)arg;
